Struct trees::tree::Tree

pub struct Tree<T> { /* private fields */ }

Composed of a root Node and a list of its child Nodes.

Implementations

impl<T> Tree<T>

pub fn new(data: T) -> Tree<T>

Creates a Tree containing only root node associated with given data.

pub fn from_tuple<Tuple, Shape>(tuple: Tuple) -> Selfwhere
Tuple: TupleTree<T, Shape>,

Constructs tree from tuple notations.

Examples

use trees::{Tree, tr};

let tree = Tree::<i32>::from_tuple(( 0, (1,2), (3,4) ));
assert_eq!( tree, tr(0) /(tr(1)/tr(2)) /(tr(3)/tr(4)) );
assert_eq!( tree.to_string(), "0( 1( 2 ) 3( 4 ) )" );

pub fn root(&self) -> &Node<T>

Reference of the root node.

pub fn root_mut(&mut self) -> Pin<&mut Node<T>>

Mutable reference of the root node.

pub fn iter_mut<'a, 's: 'a>(&'s mut self) -> IterMut<'a, T>

Provides a forward iterator over child Nodes with mutable references.

Examples

use trees::Tree;

let mut tree = Tree::<i32>::from_tuple(( 0, 1, 2, 3 ));
tree.iter_mut().for_each( |mut child| *child.data_mut() *= 10 );
assert_eq!( tree.to_string(), "0( 10 20 30 )" );

pub fn push_front(&mut self, tree: Tree<T>)

Adds the tree as the first child.

Examples

use trees::Tree;
let mut tree = Tree::new(0);
tree.push_front( Tree::new(1) );
assert_eq!( tree.to_string(), "0( 1 )" );
tree.push_front( Tree::new(2) );
assert_eq!( tree.to_string(), "0( 2 1 )" );

pub fn push_back(&mut self, tree: Tree<T>)

Adds the tree as the last child.

Examples

use trees::Tree;
let mut tree = Tree::new(0);
tree.push_back( Tree::new(1) );
assert_eq!( tree.to_string(), "0( 1 )" );
tree.push_back( Tree::new(2) );
assert_eq!( tree.to_string(), "0( 1 2 )" );

pub fn prepend(&mut self, forest: Forest<T>)

Adds all the forest’s trees at front of children list.

Examples

use trees::{Forest, Tree};

let mut tree = Tree::new(0);
tree.push_back( Tree::new(1) );
tree.push_back( Tree::new(2) );
let mut forest = Forest::new();
forest.push_back( Tree::new(3) );
forest.push_back( Tree::new(4) );
tree.prepend( forest );
assert_eq!( tree.to_string(), "0( 3 4 1 2 )" );

pub fn append(&mut self, forest: Forest<T>)

Adds all the forest’s trees at back of children list.

Examples

use trees::{Forest, Tree};

let mut tree = Tree::new(0);
tree.push_back( Tree::new(1) );
tree.push_back( Tree::new(2) );
let mut forest = Forest::new();
forest.push_back( Tree::new(3) );
forest.push_back( Tree::new(4) );
tree.root_mut().append( forest );
assert_eq!( tree.to_string(), "0( 1 2 3 4 )" );

pub fn abandon(&mut self) -> Forest<T>

Removes and returns the given Tree’s children.

Examples

use trees::{Forest, Tree};

let mut tree = Tree::new(0);
tree.push_back( Tree::new(1) );
tree.push_back( Tree::new(2) );
let forest = tree.abandon();
assert_eq!( forest.to_string(), "( 1 2 )" );
assert_eq!( tree, Tree::new(0) );

pub fn pop_front(&mut self) -> Option<Tree<T>>

Removes and returns the first child.

Examples

use trees::Tree;

let mut tree = Tree::new(0);
tree.push_back( Tree::new(1) );
tree.push_back( Tree::new(2) );
assert_eq!( tree.to_string(), "0( 1 2 )" );
assert_eq!( tree.pop_front(), Some( Tree::new(1) ));
assert_eq!( tree.to_string(), "0( 2 )" );
assert_eq!( tree.pop_front(), Some( Tree::new(2) ));
assert_eq!( tree.to_string(), "0" );
assert_eq!( tree.pop_front(), None );

pub fn pop_back(&mut self) -> Option<Tree<T>>

Removes and returns the last child.

Examples

use trees::Tree;

let mut tree = Tree::new(0);
tree.push_back( Tree::new(1) );
tree.push_back( Tree::new(2) );
assert_eq!( tree.to_string(), "0( 1 2 )" );
assert_eq!( tree.pop_back(), Some( Tree::new(2) ));
assert_eq!( tree.to_string(), "0( 1 )" );
assert_eq!( tree.pop_back(), Some( Tree::new(1) ));
assert_eq!( tree.to_string(), "0" );
assert_eq!( tree.pop_back(), None );

pub fn front_mut(&mut self) -> Option<Pin<&mut Node<T>>>

Returns a mutable reference to the first child of this node, or None if it has no child.

pub fn back_mut(&mut self) -> Option<Pin<&mut Node<T>>>

Returns a mutable reference to the last child of this node, or None if it has no child.

impl<T> Tree<T>

pub fn bfs_children_mut(&mut self) -> BfsForest<Splitted<IterMut<'_, T>>>

Provides a forward iterator with mutable references in a breadth-first manner, which iterates over all its descendants.

Examples

use trees::{tr, Tree};

let mut tree = Tree::from_tuple(( 0, (1,2,3), (4,5,6), ));
tree.bfs_children_mut().iter
    .zip( 1.. )
    .for_each( |(visit,nth)| *visit.data += 10 * nth );
assert_eq!( tree, Tree::<i32>::from_tuple(( 0, (11,32,43), (24,55,66), )));

pub fn bfs_mut(&mut self) -> BfsTree<Splitted<IterMut<'_, T>>>

Provides a forward iterator with mutable references in a breadth-first manner.

Examples

use trees::{tr, Tree};

let mut tree = Tree::from_tuple(( 0, (1,2,3), (4,5,6), ));
tree.bfs_mut().iter
    .zip( 1.. )
    .for_each( |(visit,nth)| *visit.data += 10 * nth );
assert_eq!( tree, Tree::<i32>::from_tuple(( 10, (21,42,53), (34,65,76), )));

pub fn into_bfs(self) -> BfsTree<Splitted<IntoIter<T>>>

Provides a forward iterator with owned data in a breadth-first manner.

Examples

use trees::{bfs,Size};
use trees::Tree;

let tree = Tree::<i32>::from_tuple(( 0, (1,2,3), (4,5,6) ));
let visits = tree.into_bfs().iter.collect::<Vec<_>>();
assert_eq!( visits, vec![
    bfs::Visit{ data: 0, size: Size{ degree: 2, descendants: 6 }},
    bfs::Visit{ data: 1, size: Size{ degree: 2, descendants: 2 }},
    bfs::Visit{ data: 4, size: Size{ degree: 2, descendants: 2 }},
    bfs::Visit{ data: 2, size: Size{ degree: 0, descendants: 0 }},
    bfs::Visit{ data: 3, size: Size{ degree: 0, descendants: 0 }},
    bfs::Visit{ data: 5, size: Size{ degree: 0, descendants: 0 }},
    bfs::Visit{ data: 6, size: Size{ degree: 0, descendants: 0 }},
]);