Struct trees::forest::Forest

pub struct Forest<T> { /* fields omitted */ }

List of Nodes as its children.

Implementations

impl<T> Forest<T>

pub fn new() -> Forest<T>

Makes an empty Forest.

pub fn from_tuple<Tuple, Shape>(tuple: Tuple) -> Self where
Tuple: TupleForest<T, Shape>,

Construct forest from tuple notations.

Examples

use trees::{Forest, tr};

let forest = Forest::<i32>::from_tuple(( 0, (1,2), (3,4) ));
assert_eq!( forest, -tr(0) -tr(1)/tr(2) -tr(3)/tr(4) );
assert_eq!( forest.to_string(), "( 0 1( 2 ) 3( 4 ) )" );

pub fn has_no_child(&self) -> bool

Returns true if Forest is empty.

Examples

use trees::{tr, fr};
let mut forest = fr();
assert!( forest.has_no_child() );
forest.push_back( tr(1) );
assert!( !forest.has_no_child() );

pub fn degree(&self) -> usize

Returns the number of child nodes in Forest.

Examples

use trees::Forest;
let forest = Forest::<i32>::from_tuple(( 0, (1,2), (3,4) ));
assert_eq!( forest.degree(), 3 );

pub fn node_count(&self) -> usize

Returns the number of all child nodes in Forest.

Examples

use trees::Forest;
let forest = Forest::<i32>::from_tuple(( 0, (1,2), (3,4) ));
assert_eq!( forest.node_count(), 5 );

pub fn iter<'a, 's: 'a>(&'s self) -> Iter<'a, T>

Provides a forward iterator over child Nodes.

Examples

use trees::{tr, fr};

let forest = fr::<i32>();
assert_eq!( forest.iter().next(), None );

let forest = -tr(1)-tr(2);
let mut iter = forest.iter();
assert_eq!( iter.next(), Some( tr(1).root() ));
assert_eq!( iter.next(), Some( tr(2).root() ));
assert_eq!( iter.next(), None );

pub fn iter_mut<'a, 's: 'a>(&'s mut self) -> IterMut<'a, T>

Provides a forward iterator over child Nodes with mutable references.

Examples

use trees::Forest;

let mut forest = Forest::<i32>::new();
assert_eq!( forest.iter_mut().next(), None );

let mut forest = Forest::<i32>::from_tuple(( 1, 2 ));
forest.iter_mut().for_each( |mut child| { *child.data_mut() *= 10; });
assert_eq!( forest.to_string(), "( 10 20 )" );

pub fn front(&self) -> Option<&Node<T>>

Returns the first child of the forest, or None if it is empty.

pub fn front_mut(&mut self) -> Option<Pin<&mut Node<T>>>

Returns a mutable pointer to the first child of the forest, or None if it is empty.

pub fn back(&self) -> Option<&Node<T>>

pub fn back_mut(&mut self) -> Option<Pin<&mut Node<T>>>

Returns a mutable pointer to the last child of the forest, or None if it is empty.

pub fn push_front(&mut self, tree: Tree<T>)

Add the tree as the first child.

Examples

use trees::{Tree, Forest};
let mut forest = Forest::new();
forest.push_front( Tree::new(1) );
assert_eq!( forest.to_string(), "( 1 )" );
forest.push_front( Tree::new(2) );
assert_eq!( forest.to_string(), "( 2 1 )" );

pub fn push_back(&mut self, tree: Tree<T>)

Add the tree as the last child.

Examples

use trees::{Tree, Forest};
let mut forest = Forest::new();
forest.push_back( Tree::new(1) );
assert_eq!( forest.to_string(), "( 1 )" );
forest.push_back( Tree::new(2) );
assert_eq!( forest.to_string(), "( 1 2 )" );

pub fn pop_front(&mut self) -> Option<Tree<T>>

Remove and return the first child.

Examples

use trees::{Tree, Forest};
let mut forest = Forest::new();
forest.push_back( Tree::new(1) );
forest.push_back( Tree::new(2) );
assert_eq!( forest.to_string(), "( 1 2 )" );
assert_eq!( forest.pop_front(), Some( Tree::new(1) ));
assert_eq!( forest.to_string(), "( 2 )" );
assert_eq!( forest.pop_front(), Some( Tree::new(2) ));
assert_eq!( forest.to_string(), "()" );

pub fn pop_back(&mut self) -> Option<Tree<T>>

Remove and return the first child.

Examples

use trees::{Tree, Forest};
let mut forest = Forest::new();
forest.push_back( Tree::new(1) );
forest.push_back( Tree::new(2) );
assert_eq!( forest.to_string(), "( 1 2 )" );
assert_eq!( forest.pop_back(), Some( Tree::new(2) ));
assert_eq!( forest.to_string(), "( 1 )" );
assert_eq!( forest.pop_back(), Some( Tree::new(1) ));
assert_eq!( forest.to_string(), "()" );

pub fn prepend(&mut self, forest: Forest<T>)

Add all the forest’s trees at front of children list

Examples

use trees::{Tree, Forest};
let mut forest = Forest::new();
forest.push_back( Tree::new(1) );
forest.push_back( Tree::new(2) );
let mut forest2 = Forest::new();
forest2.push_back( Tree::new(3) );
forest2.push_back( Tree::new(4) );
forest.prepend( forest2 );
assert_eq!( forest.to_string(), "( 3 4 1 2 )" );

pub fn append(&mut self, forest: Forest<T>)

Add all the forest’s trees at back of children list

Examples

use trees::{Tree, Forest};
let mut forest = Forest::new();
forest.push_back( Tree::new(1) );
forest.push_back( Tree::new(2) );
let mut forest2 = Forest::new();
forest2.push_back( Tree::new(3) );
forest2.push_back( Tree::new(4) );
forest.append( forest2 );
assert_eq!( forest.to_string(), "( 1 2 3 4 )" );

impl<T> Forest<T>

pub fn bfs(&self) -> BfsForest<Splitted<Iter<'_, T>>>

Provides a forward iterator in a breadth-first manner.

Examples

use trees::Forest;

let forest = Forest::from_tuple(( (1,2,3), (4,5,6), ));
let visits = forest.bfs().iter
    .map( |visit| (*visit.data, visit.size.degree, visit.size.descendants ))
    .collect::<Vec<_>>();
assert_eq!( visits, vec![ (1, 2, 2), (4, 2, 2), (2, 0, 0), (3, 0, 0), (5, 0, 0), (6, 0, 0), ]);

pub fn bfs_mut(&mut self) -> BfsForest<Splitted<IterMut<'_, T>>>

Provides a forward iterator with mutable references in a breadth-first manner.

Examples

use trees::Forest;

let mut forest = Forest::<i32>::from_tuple(( (1,2,3), (4,5,6), ));
forest.bfs_mut().iter
    .zip( 0.. )
    .for_each( |(visit,nth)| *visit.data += 10 * nth );
assert_eq!( forest, Forest::from_tuple(( (1,(22,),(33,)), (14,(45,),(56,)), )));

pub fn into_bfs(self: Forest<T>) -> BfsForest<Splitted<IntoIter<T>>>

Provides a forward iterator with owned data in a breadth-first manner.

Examples

use trees::{bfs,Size};
use trees::Forest;

let forest = Forest::<i32>::new();
let visits = forest.into_bfs().iter.collect::<Vec<_>>();
assert!( visits.is_empty() );

let forest = Forest::from_tuple(( (1,2,3), (4,5,6), ));
let visits = forest.into_bfs().iter.collect::<Vec<_>>();
assert_eq!( visits, vec![
    bfs::Visit{ data: 1, size: Size{ degree: 2, descendants: 2 }},
    bfs::Visit{ data: 4, size: Size{ degree: 2, descendants: 2 }},
    bfs::Visit{ data: 2, size: Size{ degree: 0, descendants: 0 }},
    bfs::Visit{ data: 3, size: Size{ degree: 0, descendants: 0 }},
    bfs::Visit{ data: 5, size: Size{ degree: 0, descendants: 0 }},
    bfs::Visit{ data: 6, size: Size{ degree: 0, descendants: 0 }},
]);